people-counting-service/requirements.txt
attrs==22.1.0
black==22.10.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
exceptiongroup==1.0.4
filelock==3.8.2
identify==2.5.9
idna==3.4
iniconfig==1.1.1
isort==5.10.1
mypy-extensions==0.4.3
nodeenv==1.7.0
numpy==1.23.5
packaging==21.3
pathspec==0.10.2
pika==1.3.1
Pillow==9.3.0
platformdirs==2.6.0
pluggy==1.0.0
pre-commit==2.20.0
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
python-decouple==3.6
PyYAML==6.0
requests==2.28.1
toml==0.10.2
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.13
virtualenv==20.17.1
